+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/GetComments.kt
@@ -8,9 +8,9 @@ import kotlinx.serialization.Serializable
 internal data class GetComments(
     val type_: ListingType? /* "All" | "Local" | "Subscribed" */ = null,
     val sort: CommentSortType? /* "Hot" | "Top" | "New" | "Old" */ = null,
-    val max_depth: Int? = null,
-    val page: Int? = null,
-    val limit: Int? = null,
+    val max_depth: Long? = null,
+    val page: Long? = null,
+    val limit: Long? = null,
     val community_id: CommunityId? = null,
     val community_name: String? = null,
     val post_id: PostId? = null,
