+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/ListPostReports.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class ListPostReports(
-    val page: Int? = null,
-    val limit: Int? = null,
+    val page: Long? = null,
+    val limit: Long? = null,
     val unresolved_only: Boolean? = null,
     val community_id: CommunityId? = null,
     val auth: String,
