+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/SiteAggregates.kt
@@ -4,14 +4,14 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class SiteAggregates(
-    val id: Int,
+    val id: Long,
     val site_id: SiteId,
-    val users: Int,
-    val posts: Int,
-    val comments: Int,
-    val communities: Int,
-    val users_active_day: Int,
-    val users_active_week: Int,
-    val users_active_month: Int,
-    val users_active_half_year: Int,
+    val users: Long,
+    val posts: Long,
+    val comments: Long,
+    val communities: Long,
+    val users_active_day: Long,
+    val users_active_week: Long,
+    val users_active_month: Long,
+    val users_active_half_year: Long,
 )
