+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/SiteAggregates.kt
@@ -5,12 +5,12 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SiteAggregates(
     val site_id: SiteId,
-    val users: Int,
-    val posts: Int,
-    val comments: Int,
-    val communities: Int,
-    val users_active_day: Int,
-    val users_active_week: Int,
-    val users_active_month: Int,
-    val users_active_half_year: Int,
+    val users: Long,
+    val posts: Long,
+    val comments: Long,
+    val communities: Long,
+    val users_active_day: Long,
+    val users_active_week: Long,
+    val users_active_month: Long,
+    val users_active_half_year: Long,
 )
