+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x19/datatypes/SiteId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x19.datatypes
 
-typealias SiteId = Int
+typealias SiteId = Long
